requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==10.0.1
promise==2.3
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
